requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
cryptography==43.0.3
Django==5.1.1
django-allauth==65.1.0
django-crispy-forms==2.3
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
psycopg2-binary==2.9.10
pycparser==2.22
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
